rename : 오늘의 운세 조회 관련 이름 변경 및 directory 위치 변경
