+++ src/test/java/com/example/classicjeans/controller/AlanBaziControllerTest.java
@@ -1,5 +1,6 @@
 package com.example.classicjeans.controller;
 
+import com.example.classicjeans.controller.rest.AlanBaziController;
 import com.example.classicjeans.dto.request.AlanBaziRequest;
 import com.example.classicjeans.dto.response.AlanBaziResponse;
 import com.example.classicjeans.service.AlanBaziService;
@@ -14,7 +15,6 @@ import org.springframework.test.web.servlet.setup.MockMvcBuilders;
 
 import java.time.LocalDate;
 
-import static org.junit.jupiter.api.Assertions.*;
 import static org.mockito.Mockito.when;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
